+++ js/settings-admin.js
@@ -44,6 +44,7 @@ window.addEventListener("DOMContentLoaded", function () {
     $("#breezedark-theme-enabled").change(function () {
         $.post(OC.generateUrl("apps/breezedark/settings/admin"), {
             theme_enabled: this.checked ? 1 : 0,
+            theme_automatic_activation_enabled: $("#breezedark-automatic-activation-enabled").prop("checked") ? 1 : 0,
             theme_login_page: $("#breezedark-theme-login-page").prop("checked") ? 1 : 0,
         })
             .done(function () {
@@ -61,11 +62,29 @@ window.addEventListener("DOMContentLoaded", function () {
             !$("#breezedark-theme-enabled").prop("checked")
         );
     });
+	
+	$("#breezedark-automatic-activation-enabled").change(function () {
+        $.post(OC.generateUrl("apps/breezedark/settings/admin"), {
+			theme_enabled: $("#breezedark-theme-enabled").prop("checked") ? 1 : 0,
+            theme_automatic_activation_enabled: this.checked ? 1 : 0,
+            theme_login_page: $("#breezedark-theme-login-page").prop("checked") ? 1 : 0,
+        })
+            .done(function () {
+                postSuccess(
+                    "label[for='breezedark-automatic-activation-enabled']",
+                    "breezedark-theme-automatic-activation-enabled-msg"
+                );
+            })
+            .fail(function () {
+                postError("label[for='breezedark-automatic-activation-enabled']", "breezedark-theme-automatic-activation-enabled-msg");
+            });
+    });
 
     $("#breezedark-theme-login-page").change(function () {
         $.post(OC.generateUrl("apps/breezedark/settings/admin"), {
             theme_login_page: this.checked ? 1 : 0,
             theme_enabled: $("#breezedark-theme-enabled").prop("checked") ? 1 : 0,
+			theme_automatic_activation_enabled: $("#breezedark-automatic-activation-enabled").prop("checked") ? 1 : 0,
         })
             .done(function () {
                 postSuccess(